requirements.txt
Django==3.1
gunicorn
django-environ
psycopg2-binary
django-storages[google]
google-cloud-secret-manager==2.0.0


# required for internal functionality, not deployement related
requests
pyyaml
bs4
